10월 24일 20시 00분 appointment 의사정보 불러오기 완료

+++ src/main/java/com/example/test/user/vo/DocVO.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 @Component
 @Data
 public class DocVO {
-    private int docNo;
+    private Long docNo;
     private String docId;
     private String docPw;
     private String docName;
